helpers: refactor date fixing and hash making so all scrapers can use
