+++ oroi-scrape/scrapers/bristol.py
@@ -1,23 +1,7 @@
 import copy
 import lxml
 
-from memorious.helpers.key import make_id
-
-
-def improve_register_date(datestring):
-    """
-    These dates are so terribly formatted that generic fuzzy date
-    parsing can't handle them. This doesn't parse them completely,
-    (because we don't want to mess with the source data too much)
-    just makes them a little less awful for future processing.
-
-    ie. "This register of interests was published on Wednesday, 22nd June, 2016, 12.16 pm."
-    """
-    datestring = datestring.replace("This register of interests was published on ", "")
-    datestring = datestring.replace(" pm.", "PM").replace(" pm", "PM")
-    datestring = datestring.replace(" am.", "AM").replace(" am", "AM")
-    datestring = datestring.replace(".", ":")
-    return datestring
+from scrapers.helpers import improve_register_date, make_hashes
 
 
 def get_register(context, data):
@@ -76,7 +60,6 @@ def parse_register(context, data):
     with context.http.rehash(data) as result:
         output_base = {
             "source": result.url,
-            "source_id": make_id(result.url, data.get("member_name")),
             "member_name": data.get("member_name"),
             "member_url": data.get("member_url"),
             "declared_to": "Bristol City Council",
@@ -88,9 +71,6 @@ def parse_register(context, data):
             bullets = holder.findall(".//div[@class='mgLinks']//li")
             declared_date = improve_register_date(bullets[0].text)
             output_base["declared_date"] = declared_date
-            output_base["registration_id"] = make_id(
-                result.url, data.get("member_name"), output_base.get("declared_date")
-            )
 
             content = holder.find(".//div[@class='mgDeclarations']")
             declarations = content.findall(".//table")
@@ -101,11 +81,7 @@ def parse_register(context, data):
                         output = copy.deepcopy(output_base)
 
                         output["interest_type"] = declaration_mapping.get(number)
-                        output["declaration_id"] = make_id(
-                            result.url,
-                            data.get("member_name"),
-                            output.get("interest_type"),
-                        )
+
                         if notes_mapping.get(number) is not None:
                             output["notes"] = notes_mapping.get(number)
 
@@ -118,14 +94,7 @@ def parse_register(context, data):
                                     output["description"] = cells[0].text_content()
                                     output["interest_date"] = cells[1].text_content()
 
-                                    output["interest_hash"] = make_id(
-                                        output.get("interest_type"),
-                                        output.get("description"),
-                                        output.get("interest_date"),
-                                        output.get("interest_from"),
-                                        output.get("member_name"),
-                                    )
-
+                                    output = make_hashes(output)
                                     context.emit(rule="store", data=output)
 
                         else:
@@ -138,14 +107,8 @@ def parse_register(context, data):
                                     .replace("\n", "")
                                     .strip()
                                 )
-                                output["interest_hash"] = make_id(
-                                    output.get("interest_type"),
-                                    output.get("description"),
-                                    output.get("interest_date"),
-                                    output.get("interest_from"),
-                                    output.get("member_name"),
-                                )
 
+                                output = make_hashes(output)
                                 context.emit(rule="store", data=output)
 
 
@@ -182,22 +145,15 @@ def parse_meetings(context, data):
 
                     output = {
                         "source": meeting_link,
-                        "source_id": make_id(meeting_link, member_name),
                         "declared_date": meeting_date,
                         "declared_to": "Bristol City Council ({})".format(
                             meeting_group
                         ),
                         "member_name": member_name.strip(),
                         "member_url": member_url,
-                        "registration_id": make_id(
-                            meeting_link, member_name, meeting_date
-                        ),
-                        "declaration_id": make_id(
-                            meeting_link, member_name, "positions"
-                        ),
                         "interest_type": "positions",
                         "description": description,
-                        "interest_hash": make_id("positions", description, member_name),
                     }
 
+                    output = make_hashes(output)
                     context.emit(data=output)
